Varios ajustes

+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/models/supports/Operation.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 /**
  * @author Paulo Weskley de Almeida Ferreira
  * @date 2020-07-27
- * <p>
+ *
  * Classe Responsável por mapear a tabela auxiliar das operações de pedidos de estoque.
  */
 
@@ -19,12 +19,12 @@ public class Operation implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String description;
 
     /*Constructors*/
-    public Operation(Long id, String description) {
+    public Operation(Integer id, String description) {
         this.id = id;
         this.description = description;
     }
@@ -33,11 +33,11 @@ public class Operation implements Serializable {
     }
 
     /*Getters and Setters*/
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
