+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/repositories/StockOrderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockOrderRepository extends JpaRepository<StockOrder, Long> {
+public interface StockOrderRepository extends JpaRepository<StockOrder, Integer> {
 }
