+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/repositories/SubsidiaryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubsidiaryRepository extends JpaRepository<Subsidiary, Long> {
+public interface SubsidiaryRepository extends JpaRepository<Subsidiary, Integer> {
     Subsidiary getByCnpj(String code);
 }
