+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/services/StockService.java
@@ -17,7 +17,7 @@ public class StockService {
         return this.repository.findAll();
     }
 
-    public Stock findById(Long id) {
+    public Stock findById(Integer id) {
         return this.repository.findById(id).get();
     }
 }
