+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/services/SubsidiaryService.java
@@ -17,7 +17,7 @@ public class SubsidiaryService {
         return this.repository.findAll();
     }
 
-    public Subsidiary findById(Long id) {
+    public Subsidiary findById(Integer id) {
         return this.repository.findById(id).get();
     }
