+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/services/exceptions/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package br.com.pwneo.estoque_back_end.services.exceptions;
 public class ResourceNotFoundException extends RuntimeException{
     private static final Long serialVersionUID = 1L;
 
-    public ResourceNotFoundException(Long id){
+    public ResourceNotFoundException(Integer id){
         super("Resource not found. Id " + id);
     }
 }
