+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/services/supports/OperationService.java
@@ -17,7 +17,7 @@ public class OperationService {
         return this.repository.findAll();
     }
 
-    public Operation findById(Long id) {
+    public Operation findById(Integer id) {
         return this.repository.findById(id).get();
     }
 }
