+++ implementacao/estoque_back_end/src/main/java/br/com/pwneo/estoque_back_end/services/supports/PaymentMethodService.java
@@ -17,7 +17,7 @@ public class PaymentMethodService {
         return this.repository.findAll();
     }
 
-    public PaymentMethod findById(Long id) {
+    public PaymentMethod findById(Integer id) {
         return this.repository.findById(id).get();
     }
 }
